batch_only/gcs_to_bq/requirements.txt
# Google Cloud libraries for BigQuery and Cloud Storage
google-cloud-bigquery>=3.11.0
google-cloud-storage>=2.10.0

# Core dependencies
pandas>=2.0.0
pyarrow>=12.0.0
pandas-gbq>=0.26.1

# Optional: For enhanced functionality
google-cloud-bigquery-storage>=2.20.0  # For faster BigQuery I/O

# Authentication and credentials
google-auth>=2.40.3
python-dotenv>=1.0.0
